refactor: renamed `Barcode` to `IOBarcode` and hooks
