+++ ts/features/idpay/payment/components/Barcode/IOBarcode.ts
