+++ ts/features/idpay/payment/screens/IDPayPaymentCodeScanScreen.tsx
@@ -15,10 +15,10 @@ import {
   IOStackNavigationProp
 } from "../../../../navigation/params/AppParamsList";
 import { openWebUrl } from "../../../../utils/url";
-import { useIOBarcodeScanner } from "../components/Barcode/BarcodeScanner";
+import { useIOBarcodeScanner } from "../components/Barcode/useIOBarcodeScanner";
 import { BottomTabNavigation } from "../components/BottomTabNavigation";
 import { CameraPermissionView } from "../components/CameraPermissionView";
-import { IOBarcode } from "../components/Barcode/Barcode";
+import { IOBarcode } from "../components/Barcode/IOBarcode";
 
 const IDPayPaymentCodeScanScreen = () => {
   const isFocused = useIsFocused();
